Fail link to an installed-but-undocumented module
This is a follow-on to the Pod::Perldoc test failure noted in the
previous commit. It was passing when I ran it because the missing
documentation caused the logic to assume the module was not installed,
and execute the "does the module exist" logic. What I think happened
then was that it passed on my system, but failed on the smoker because
Andreas J. König (ANDK) does not allow net access.

After some thought I decided that failure was the correct behavior in
this case.