Report actual line numbers of links (I hope)
This involved changing out the parser. Instead of using ::SimpleTree and
rummaging through the parse, I used ::PullParser and collected
information as I went. With preserve_whitespace() set true, I can infer
the line number of the link from the starting line of the paragraph it
is embedded in by counting line breaks.