Fix test for link to non-existent man page.
It turns out my somple-minded attempt to come up with a non-existent man
page ('bad man') did not work under FreeBSD because some component (man
itself?) split the quoted argument 'bad man' anyway. 'bad' was not a
valid man page, but 'man' of course was, so the status was 0 and the
test failed -- or rather, it passed when it should have failed. Sigh.

Version 0.000_905 appears not to have this problem, but looks are
deceiving. It will skip this test because it executes 'man -w 1 bad\
man' on its own, and since that passes the "bad man" test gets skipped.
Sigh again.