Fix test of link to extant-but-undocumented module.
Occasionally there is a module that exists but whose documentation is
somewhere other than the module itself or the same-named .pod file.
XML::Twig::XPath is (currently) an example of this, with the docs
residing in XML::Twig.

The prior incorrect behavior was to report a failure but omit the link
and line number, and then generate a passing test. This commit instates
the correct behavior -- report a failure on (e.g.) L<XML::Twig::XPath>.