Add canned 'ignore_url' value MAYBE_IGNORE_GITHUB.
This causes GitHub URLs to be ignored unless the source directory
contains a Git repository and GitHub is configured as a remote.  This is
intended to be useful to authors whose boiler plate contains GitHub
links, in the early stages of the project before the GitHub repository
has been created.