Fix bug in add_dir functionality.
I was originally just slapping it onto the front of @INC. But
Pod::Perldoc takes the Perl bin directory before anything in @INC, and
so was finding the installed version of the script/ contents rather than
the one in the development directory.

The solution was to implement my own search through the add_dir
directories. If this gets a hit, Pod::Perldoc is never used.